Refactored code so now a model is created in models folder to represent a state object for a particular object. This keeps the model and its actions in one file per model.  Created utils for AVFormsActions and AVFormReducers to allow multiple usage of validations.
